core/standards/web-apps/media/semantics/muted/001.html

<!doctype html>
<html>
<head>
<title>&lt;audio muted> + &lt;video muted></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<audio></audio><audio muted></audio><audio muted="false"></audio>
<video></video><video muted></video><video muted="false"></video>
<script>
['audio', 'video'].forEach(function(tagName) {
  test(function() {
    var e = document.querySelector(tagName);
    assert_false(e.muted);
    assert_false(e.defaultMuted);
  }, '<'+tagName+'>');

  test(function() {
    var e = document.querySelector(tagName+'[muted]');
    assert_true(e.muted);
    assert_true(e.defaultMuted);
  }, '<'+tagName+' muted>');

  test(function() {
    var e = document.querySelector(tagName+'[muted="false"]');
    assert_true(e.muted);
    assert_true(e.defaultMuted);
  }, '<'+tagName+' muted="false">');

  test(function() {
    var e = document.createElement(tagName);
    assert_equals(typeof e.muted, "boolean");
    assert_equals(typeof e.defaultMuted, "boolean");
  }, tagName+'.muted and .defaultMuted type');

  test(function() {
    var e = document.createElement(tagName);
    e.setAttribute('muted', '');
    assert_false(e.muted);
    assert_true(e.defaultMuted);
  }, tagName+'.setAttribute("muted", "")');

  test(function() {
    var e = document.createElement(tagName);
    e.defaultMuted = true;
    e.removeAttribute('muted');
    assert_false(e.muted);
    assert_false(e.defaultMuted);
  }, tagName+'.removeAttribute("muted")');

  test(function() {
    var e = document.createElement(tagName);
    e.defaultMuted = true;
    assert_true(e.hasAttribute('muted'));
    e.defaultMuted = false;
    assert_false(e.hasAttribute('muted'));
  }, tagName+'.defaultMuted reflection');

  test(function() {
    var e = document.createElement(tagName);
    e.muted = true;
    assert_false(e.hasAttribute('muted'));
    e.setAttribute('muted', '');
    e.muted = false;
    assert_true(e.hasAttribute('muted'));
  }, tagName+'.muted (lack of) reflection');
});
</script>
</body>
</html>
